@@ -1,35 +1,33 @@
 module github.com/DelineaXPM/dsv-github-action
 
-go 1.19
+go 1.21
 
 require (
-	github.com/bitfield/script v0.21.4
-	github.com/caarlos0/env/v6 v6.10.1
-	github.com/magefile/mage v1.14.0
+	github.com/bitfield/script v0.22.0
+	github.com/caarlos0/env/v10 v10.0.0
+	github.com/magefile/mage v1.15.0
 	github.com/matryer/is v1.4.1
-	github.com/pterm/pterm v0.12.56
-	github.com/sheldonhull/magetools v1.0.0
+	github.com/pterm/pterm v0.12.75
+	github.com/sheldonhull/magetools v1.0.1
 )
 
-require github.com/google/go-cmp v0.5.8 // indirect
-
 require (
-	atomicgo.dev/cursor v0.1.1 // indirect
+	atomicgo.dev/cursor v0.2.0 // indirect
 	atomicgo.dev/keyboard v0.2.9 // indirect
-	bitbucket.org/creachadair/shell v0.0.7 // indirect
+	atomicgo.dev/schedule v0.1.0 // indirect
 	github.com/containerd/console v1.0.3 // indirect
-	github.com/dustin/go-humanize v1.0.0 // indirect
-	github.com/gookit/color v1.5.2 // indirect
-	github.com/itchyny/gojq v0.12.7 // indirect
-	github.com/itchyny/timefmt-go v0.1.3 // indirect
-	github.com/lithammer/fuzzysearch v1.1.5 // indirect
-	github.com/logrusorgru/aurora v2.0.3+incompatible // indirect
-	github.com/mattn/go-runewidth v0.0.14 // indirect
+	github.com/dustin/go-humanize v1.0.1 // indirect
+	github.com/gookit/color v1.5.4 // indirect
+	github.com/itchyny/gojq v0.12.14 // indirect
+	github.com/itchyny/timefmt-go v0.1.5 // indirect
+	github.com/lithammer/fuzzysearch v1.1.8 // indirect
+	github.com/mattn/go-runewidth v0.0.15 // indirect
 	github.com/rivo/uniseg v0.4.4 // indirect
 	github.com/xo/terminfo v0.0.0-20220910002029-abceb7e1c41e // indirect
-	github.com/ztrue/tracerr v0.3.0 // indirect
-	golang.org/x/mod v0.8.0 // indirect
-	golang.org/x/sys v0.6.0 // indirect
-	golang.org/x/term v0.5.0 // indirect
-	golang.org/x/text v0.8.0 // indirect
+	github.com/ztrue/tracerr v0.4.0 // indirect
+	golang.org/x/mod v0.14.0 // indirect
+	golang.org/x/sys v0.16.0 // indirect
+	golang.org/x/term v0.16.0 // indirect
+	golang.org/x/text v0.14.0 // indirect
+	mvdan.cc/sh/v3 v3.7.0 // indirect
 )
